requirements.txt
pillow
rembg
flask
